Add ExtractorChrome to contrib
This adds a link extractor that uses a headless browser via the Chrome
Devtools Protocol. It works with Chromium and Google Chrome and
while untested likely also other Blink-based browsers such as Brave and
Microsoft Edge. It also works with Firefox which implements a subset of
the protocol.

In addition to ExtractorChrome itself this adds a couple of classes
under org.archive.net.chrome for communicating with the browser:

* ChromeProcess - manages the lifecycle of the browser child process
* ChromeClient  - client for the Chrome Devtools Protocol
* ChromeWindow  - client for controlling a particular browser window

Why this extractor when we already have Umbra? The primary goal is to
have something that works out of the box without any additional
infrastructure beyond having a browser installed. A longer term goal is
to be able to take advantage of tighter integration between Heritrix
and the browser to do things like recording browser subrequests to WARC
files or running ExtractorHTML against the loaded DOM.

While this extractor is already usable it is currently quite bare bones
and only extracts links from <a> and <area> tags. The intent is to
expand it with more functionality over time with features like
intercepting subrequests, reusing previously saved resources, taking
screenshots or running JavaScript behaviours.