ExtractorChrome: Capture requests made by the browser
This adds a `captureRequests` flag to ExtractorChrome which is enabled
by default and causes requests made by the browser to be captured via
the devtools Network domain. Captured browser requests are sent to the
disposition chain for WARC writing and also to statistics tracker and
crawl log.

Browser requests are given the annotation "browser" so they can be
easily distinguished in the log from normal requests.

There are a quite a few limitations that will be addressed in followup
work:

* The frontier is entirely unaware of browser requests. This means they
  bypass quotas, ignore scope rules, politeness and don't count towards
  the statistics tracked by the frontier itself.

* There's no replay of previously saved resources so duplicate requests
  for the same URL end up being made.

* Heritrix's extractors do not currently process browser requests.

* Various error and failure cases likely need improving.