Access builds.archive.org over HTTPS by using the Wayback Machine
Newer versions of Maven block access to non-HTTPS repositories in order
to provide protection against man in the middle attacks (see #310).
Unfortunately builds.archive.org does not provide public HTTPS access.

However the Wayback Machine is available over HTTPS and has the needed
artifacts (javaswf, kryo, minlog, reflectasm) archived. This is little
oddly circular but solves the immediate problem.

I changed the order of the repositories and added central explicitly so
that Maven will only fallback to Wayback for the artifacts that aren't
in the first two.

I also turned off snapshots. It looks like builds.archive.org has
stopped generating new Heritrix snapshots anyway - I don't have
access to the IA Jenkins to see why but based on the timing I'd guess it
might be because it was building with Java 7.