Enable the -parameters compiler option
This is needed for Spring 6.1 to be able to reflect on parameter names. We have some autowiring that relies on parameter names such as `AbstractFrontier.setScope(DecideRule scope)`. Without this Spring doesn't know whether to inject the `scope` or `acceptSurts` beans as they both subclass DecideRule. Apparently older versions of Spring used to discover method names by parsing bytecode but this was mechanism removed in 6.1.